plugin/app/api/ocr/requirements.txt
Flask==3.0.0
google-cloud-storage==2.7.0
pytesseract==0.3.10
Pillow==10.0.1
python-dateutil==2.8.2